# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Plataforma de Vitrines Digitais API
  description: API REST Multi-Tenant para Vitrines Digitais (Público e Painel Master).
  version: 1.0.0
servers:
  - url: /api
    description: Base API Path
paths:
...
  /admin/settings:
    get:
      summary: Obter configurações da loja do tenant
      tags:
        - Admin - Settings
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Configurações da loja
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreSettingsResponse'
    put:
      summary: Atualizar configurações da loja
      tags:
        - Admin - Settings
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStoreSettingsRequest'
      responses:
        '200':
          description: Configurações atualizadas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreSettingsResponse'
...
  /admin/testimonials:
    get:
      summary: Listar depoimentos do tenant
      tags:
        - Admin - Testimonials
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de depoimentos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TestimonialResponse'
    post:
      summary: Criar depoimento
      tags:
        - Admin - Testimonials
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestimonialInput'
      responses:
        '201':
          description: Depoimento criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestimonialResponse'
  /admin/testimonials/{id}:
    put:
      summary: Atualizar depoimento
      tags:
        - Admin - Testimonials
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestimonialInput'
      responses:
        '200':
          description: Depoimento atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestimonialResponse'
    delete:
      summary: Excluir depoimento
      tags:
        - Admin - Testimonials
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Depoimento excluído
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    StoreSettingsResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        storeName:
          type: string
        slug:
          type: string
        logoUrl:
          type: string
          nullable: true
        primaryColor:
          type: string
          nullable: true
        secondaryColor:
          type: string
          nullable: true
        backgroundColor:
          type: string
          nullable: true
        textColor:
          type: string
          nullable: true
        whatsappNumber:
          type: string
          nullable: true
        whatsappMessageTemplate:
          type: string
          nullable: true
    UpdateStoreSettingsRequest:
      type: object
      properties:
        storeName:
          type: string
        logoUrl:
          type: string
        primaryColor:
          type: string
        secondaryColor:
          type: string
        backgroundColor:
          type: string
        textColor:
          type: string
        whatsappNumber:
          type: string
        whatsappMessageTemplate:
          type: string
...
    TestimonialResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        text:
          type: string
        displayOrder:
          type: integer
        active:
          type: boolean
    TestimonialInput:
      type: object
      required:
        - name
        - text
      properties:
        name:
          type: string
        text:
          type: string
        displayOrder:
          type: integer
        active:
          type: boolean
...